- source code cleanup - organization of source files - new directory structure
