+++ app/src/main/java/com/example/deepak/simpletodo/activities/NewOrEditItemActivity.java
@@ -1,11 +1,10 @@
-package com.example.deepak.simpletodo;
+package com.example.deepak.simpletodo.activities;
 
 import android.app.DialogFragment;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
-import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
@@ -13,8 +12,11 @@ import android.widget.ArrayAdapter;
 import android.widget.EditText;
 import android.widget.Spinner;
 import android.widget.TextView;
+import com.example.deepak.simpletodo.fragments.DatePickerFragment;
+import com.example.deepak.simpletodo.utils.Item;
+import com.example.deepak.simpletodo.R;
 
-public class NewItem extends AppCompatActivity {
+public class NewOrEditItemActivity extends AppCompatActivity {
 
     Toolbar toolbar;
     Spinner priorityDropdown, statusDropdown;
