+++ app/src/main/java/com/example/deepak/simpletodo/activities/SplashScreenActivity.java
@@ -1,10 +1,11 @@
-package com.example.deepak.simpletodo;
+package com.example.deepak.simpletodo.activities;
 
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
+import com.example.deepak.simpletodo.R;
 
-public class SplashScreen extends Activity {
+public class SplashScreenActivity extends Activity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -19,7 +20,7 @@ public class SplashScreen extends Activity {
                 }catch(InterruptedException e){
                     e.printStackTrace();
                 }finally{
-                    Intent intent = new Intent(SplashScreen.this,MainActivity.class);
+                    Intent intent = new Intent(SplashScreenActivity.this,MainActivity.class);
                     startActivity(intent);
                 }
             }
