+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.example.deepak.simpletodo.MainActivity">
+    tools:context=".activities.MainActivity">
 
         <include
            android:id="@+id/tool_bar"
