+++ app/src/main/res/layout/content_main.xml
@@ -7,7 +7,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.deepak.simpletodo.MainActivity">
+    tools:context=".activities.MainActivity">
 
     <ListView
         android:layout_width="wrap_content"
