+++ app/src/main/res/menu/menu_edit.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.example.deepak.simpletodo.ItemDetails">
+    tools:context=".activities.ItemDetailsActivity">
     <item
         android:id="@+id/action_edit"
         android:orderInCategory="200"
