+++ app/src/main/res/menu/menu_new.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.example.deepak.simpletodo.NewItem">
+    tools:context=".activities.NewOrEditItemActivity">
     <item
         android:id="@+id/action_save"
         android:orderInCategory="200"
